Revision 5579
Got all HIV interventions to build as separate EMODules under SCons.
Didn't tackle much of the underlying design choices that lead to needing
to compile HIVSimpleDiagnostic.cpp and Diagnostic.cpp  many times for
this approach. Also changed auto-config on Visual Studio version for
SCons -- which was messing things up on systems with multiple VS
installs where the most-recent was not the most-wanted. Now hardcoded to
version 11.0.